matrix/requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyBluez==0.22
RPi.GPIO==0.6.3
Werkzeug==0.12.2
